(#97) refactoring domain package
