+++ storage/src/main/java/org/zalando/apidiscovery/storage/api/domain/model/ApiDto.java
@@ -1,11 +1,10 @@
-package org.zalando.apidiscovery.storage.api.domain.dto;
+package org.zalando.apidiscovery.storage.api.domain.model;
 
 import java.util.List;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.zalando.apidiscovery.storage.api.domain.ApiLifecycleState;
 
 @Data
 @NoArgsConstructor
