+++ storage/src/main/java/org/zalando/apidiscovery/storage/api/domain/model/ApiLifecycleState.java
@@ -1,4 +1,4 @@
-package org.zalando.apidiscovery.storage.api.domain;
+package org.zalando.apidiscovery.storage.api.domain.model;
 
 public enum ApiLifecycleState {
     ACTIVE,
