+++ storage/src/main/java/org/zalando/apidiscovery/storage/api/domain/model/ApplicationDto.java
@@ -1,4 +1,4 @@
-package org.zalando.apidiscovery.storage.api.domain.dto;
+package org.zalando.apidiscovery.storage.api.domain.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
