+++ storage/src/main/java/org/zalando/apidiscovery/storage/api/domain/model/VersionsDto.java
@@ -1,4 +1,4 @@
-package org.zalando.apidiscovery.storage.api.domain.dto;
+package org.zalando.apidiscovery.storage.api.domain.model;
 
 import java.util.List;
 
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.zalando.apidiscovery.storage.api.domain.ApiLifecycleState;
 
 @Data
 @Builder
