+++ storage/src/test/java/org/zalando/apidiscovery/storage/api/DomainObjectGen.java
@@ -2,7 +2,7 @@ package org.zalando.apidiscovery.storage.api;
 
 import org.zalando.apidiscovery.storage.api.repository.ApiDeploymentEntity;
 import org.zalando.apidiscovery.storage.api.repository.ApiEntity;
-import org.zalando.apidiscovery.storage.api.domain.ApiLifecycleState;
+import org.zalando.apidiscovery.storage.api.domain.model.ApiLifecycleState;
 import org.zalando.apidiscovery.storage.api.repository.ApplicationEntity;
 
 import java.time.OffsetDateTime;
