+++ storage/src/test/java/org/zalando/apidiscovery/storage/api/TestDataHelper.java
@@ -1,7 +1,7 @@
 package org.zalando.apidiscovery.storage.api;
 
 import org.springframework.core.io.Resource;
-import org.zalando.apidiscovery.storage.api.domain.dto.DiscoveredApiDefinitionDto;
+import org.zalando.apidiscovery.storage.api.domain.model.DiscoveredApiDefinitionDto;
 
 import java.io.IOException;
 import java.nio.file.Files;
